Renaming and organizing files
Summary: After doing some reading, it seems like the build backend doesn't really matter
for a package as simple as this. We have no specific build requirements and I don't see
us having any in the visible future (years to come...), so I just used hatcling, which
was the one in the packaging tutorial. Generally keeping things as simple as possible.